Extract RxVar to rx package
